Veresiye ve gramajlı ürün desteği eklendi, UI yenilendi
- Veresiye defteri, müşteri ve toptan satış modelleri eklendi
- Sepet ve satışlarda decimal (küsuratlı) adet desteği sağlandı
- Terazi barkodu okuma ve gramajlı ürün işlemleri geliştirildi
- Kullanıcı şifre hashleme kaldırıldı, düz metin olarak kaydediliyor
- Ürünlere stok takibi (StokMiktari) eklendi, satışta stok düşüyor
- Sepet ve ödeme ekranına veresiye paneli ve yeni butonlar eklendi
- AnaSayfa ve diğer sayfalarda mağaza adı ve metinler güncellendi
- Kod ve XAML dosyalarında sadeleştirme ve temizlik yapıldı
- 1 yıldan eski arşiv satışları otomatik temizleniyor
- Tüm yeni özellikler için gerekli veri ve arayüz bağlantıları sağlandı

+++ Models/ArsivedSatis.cs
@@ -6,28 +6,34 @@ namespace Saller_System.Models
     public class ArsivedSatis
     {
         [PrimaryKey, AutoIncrement]
-        public int     Id          { get; set; }
-        public int     OrijinalId  { get; set; }
-        public string  UrunAd      { get; set; } = "";
-        public int     Adet        { get; set; }
-        public decimal Fiyat       { get; set; }
-        public decimal Kar         { get; set; }
-        public DateTime Tarih      { get; set; }
-        public string? KasiyerAd   { get; set; }
-        public DateTime ArsivedAt  { get; set; } = DateTime.Now;
+        public int Id { get; set; }
+        public int OrijinalId { get; set; }
+        public string UrunAd { get; set; } = "";
+
+        // HATA ÇÖZÜMÜ: int yerine decimal yapıldı (Terazi gramajları için)
+        public decimal Adet { get; set; }
+
+        public decimal Fiyat { get; set; }
+        public decimal Kar { get; set; }
+        public DateTime Tarih { get; set; }
+        public string? KasiyerAd { get; set; }
+        public DateTime ArsivedAt { get; set; } = DateTime.Now;
 
         public ArsivedSatis() { }
 
         public ArsivedSatis(Satis s)
         {
             OrijinalId = s.Id;
-            UrunAd     = s.UrunAd ?? "";
-            Adet       = s.Adet;
-            Fiyat      = s.Fiyat;
-            Kar        = s.Kar;
-            Tarih      = s.Tarih;
-            KasiyerAd  = s.KasiyerAd;
-            ArsivedAt  = DateTime.Now;
+            UrunAd = s.UrunAd ?? "";
+
+            // Buradaki atama artık hata vermeyecek
+            Adet = s.Adet;
+
+            Fiyat = s.Fiyat;
+            Kar = s.Kar;
+            Tarih = s.Tarih;
+            KasiyerAd = s.KasiyerAd;
+            ArsivedAt = DateTime.Now;
         }
     }
 }
